reducing to a loop settings alerter
